refactor: Service 폴더로 분리 및 Error, 프로토콜을 개별 파일로 이관
